+++ src/main/java/com/example/umc8th/article/dto/ArticleResponseDTO.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 
 public class ArticleResponseDTO {
