+++ src/main/java/com/example/umc8th/article/service/command/ArticleCommandService.java
@@ -1,6 +1,5 @@
 package com.example.umc8th.article.service.command;
 
-import com.example.umc8th.article.dto.ArticleRequestDTO;
 import com.example.umc8th.article.dto.ArticleResponseDTO;
 import com.example.umc8th.article.entity.Article;
