+++ src/main/java/com/example/umc8th/article/service/command/ArticleCommandServiceImpl.java
@@ -1,6 +1,5 @@
 package com.example.umc8th.article.service.command;
 
-import com.example.umc8th.article.dto.ArticleRequestDTO;
 import com.example.umc8th.article.dto.ArticleResponseDTO;
 import com.example.umc8th.article.entity.Article;
 import com.example.umc8th.article.repository.ArticleRepository;
