+++ src/main/java/com/example/umc8th/article/service/query/ArticleQueryService.java
@@ -1,8 +1,6 @@
 package com.example.umc8th.article.service.query;
 
-import com.example.umc8th.article.dto.ArticleRequestDTO;
 import com.example.umc8th.article.dto.ArticleResponseDTO;
-import com.example.umc8th.article.entity.Article;
 
 import java.util.List;
