+++ src/main/java/com/example/umc8th/global/Entity/BaseEntity.java
@@ -1,6 +1,8 @@
 package com.example.umc8th.global.Entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
