+++ src/main/java/com/example/umc8th/reply/repository/ReplyRepository.java
@@ -1,7 +1,6 @@
 package com.example.umc8th.reply.repository;
 
 import com.example.umc8th.article.entity.Article;
-import com.example.umc8th.reply.dto.ReplyResponseDTO;
 import com.example.umc8th.reply.entity.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
