+++ src/main/java/com/example/umc8th/reply/service/command/ReplyCommandService.java
@@ -1,8 +1,5 @@
 package com.example.umc8th.reply.service.command;
 
-
-import com.example.umc8th.article.entity.Article;
-import com.example.umc8th.reply.dto.ReplyRequestDTO;
 import com.example.umc8th.reply.dto.ReplyResponseDTO;
 import com.example.umc8th.reply.entity.Reply;
