+++ src/main/java/com/example/umc8th/reply/service/query/ReplyQueryService.java
@@ -1,7 +1,6 @@
 package com.example.umc8th.reply.service.query;
 
 import com.example.umc8th.reply.dto.ReplyResponseDTO;
-import com.example.umc8th.reply.entity.Reply;
 
 import java.util.List;
