+++ src/main/java/com/example/umc8th/reply/service/query/ReplyQueryServiceImpl.java
@@ -1,9 +1,9 @@
 package com.example.umc8th.reply.service.query;
 
 import com.example.umc8th.article.entity.Article;
+import com.example.umc8th.article.exception.ArticleException;
+import com.example.umc8th.article.exception.code.ArticleErrorCode;
 import com.example.umc8th.article.repository.ArticleRepository;
-import com.example.umc8th.global.apiPayload.code.ArticleErrorCode;
-import com.example.umc8th.global.apiPayload.exception.ArticleException;
 import com.example.umc8th.reply.dto.ReplyResponseDTO;
 import com.example.umc8th.reply.entity.Reply;
 import com.example.umc8th.reply.repository.ReplyRepository;
@@ -25,7 +25,7 @@ public class ReplyQueryServiceImpl implements ReplyQueryService {
     public List<ReplyResponseDTO.ReplyDTO> getReplyList(Long articleId) {
         Article article = articleRepository.findById(articleId).orElseThrow(() ->
                 new ArticleException(ArticleErrorCode.NOT_FOUND_404));
-        List<Reply> replies = replyRepository.findAllByArticle((article));
+        List<Reply> replies = replyRepository.findAllByArticle(article);
         List<ReplyResponseDTO.ReplyDTO> replyList = new ArrayList<>();
         for (Reply reply : replies) {
             replyList.add(ReplyResponseDTO.ReplyDTO.toDTO(reply));
